chore: refactor home view into lib
