+++ apps/design-system/.storybook/tsconfig.json
@@ -3,9 +3,7 @@
   "compilerOptions": {
     "emitDecoratorMetadata": true
   },
-  "exclude": [
-    "../**/*.spec.ts"
-  ],
+  "exclude": ["../**/*.spec.ts"],
   "include": [
     "../src/**/*.stories.ts",
     "../src/**/*.stories.js",
