+++ libs/waweb-ui/src/index.ts
@@ -1,2 +1,4 @@
-export { SocialIcon } from './lib/gfx/social-icon';
-export type { SocialIconProps, SocialVariant } from './lib/gfx/social-icon';
+export { SocialIcon } from './ui/social-icon';
+export type { SocialIconProps, SocialVariant } from './ui/social-icon';
+export { HomeView } from './ui/views/home-view';
+export type { HomeViewProps } from './ui/views/home-view';
