+++ libs/waweb-ui/src/ui/social-icon/social-icon.types.ts
