+++ libs/waweb-ui/src/ui/views/home-view/home-view.fixtures.ts
@@ -1,0 +1,207 @@
+import {
+  ArrowPathIcon,
+  CloudArrowUpIcon,
+  LockClosedIcon,
+  PhoneIcon,
+  PlayCircleIcon,
+  RectangleGroupIcon,
+  ServerIcon,
+} from '@heroicons/react/20/solid';
+import {
+  ChartPieIcon,
+  CursorArrowRaysIcon,
+  FingerPrintIcon,
+  SquaresPlusIcon,
+} from '@heroicons/react/24/outline';
+import { SocialVariant } from '../../social-icon/social-icon.types';
+
+export const solutions = [
+  {
+    name: 'Design System',
+    description: 'Get a better understanding where your traffic is coming from',
+    href: '#',
+    icon: ChartPieIcon,
+  },
+  {
+    name: 'Site Builder',
+    description: 'Speak directly to your customers with our engagement tool',
+    href: '#',
+    icon: CursorArrowRaysIcon,
+  },
+  {
+    name: 'Bean Counter',
+    description: 'Your customers’ data will be safe and secure',
+    href: '#',
+    icon: FingerPrintIcon,
+  },
+  {
+    name: 'Auto DevOps',
+    description: 'Your customers’ data will be safe and secure',
+    href: '#',
+    icon: SquaresPlusIcon,
+  },
+];
+
+export const callsToAction = [
+  { name: 'See demo', href: '/demo', icon: PlayCircleIcon },
+  { name: 'Contact us', href: '/contact', icon: PhoneIcon },
+  { name: 'View all', href: '/solutions', icon: RectangleGroupIcon },
+];
+
+export const highlights = [
+  {
+    name: 'Push to deploy.',
+    description: 'Lorem ipsum, dolor sit amet consectetur adipisicing elit aute id magna.',
+    icon: CloudArrowUpIcon,
+  },
+  {
+    name: 'SSL certificates.',
+    description:
+      'Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui lorem cupidatat commodo.',
+    icon: LockClosedIcon,
+  },
+  {
+    name: 'Database backups.',
+    description: 'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus.',
+    icon: ServerIcon,
+  },
+];
+
+export const features = [
+  {
+    name: 'Push to deploy',
+    description:
+      'Commodo nec sagittis tortor mauris sed. Turpis tortor quis scelerisque diam id accumsan nullam tempus. Pulvinar etiam lacus volutpat eu. Phasellus praesent ligula sit faucibus.',
+    href: '#',
+    icon: CloudArrowUpIcon,
+  },
+  {
+    name: 'SSL certificates',
+    description:
+      'Pellentesque enim a commodo malesuada turpis eleifend risus. Facilisis donec placerat sapien consequat tempor fermentum nibh.',
+    href: '#',
+    icon: LockClosedIcon,
+  },
+  {
+    name: 'Simple queues',
+    description:
+      'Pellentesque sit elit congue ante nec amet. Dolor aenean curabitur viverra suspendisse iaculis eget. Nec mollis placerat ultricies euismod ut condimentum.',
+    href: '#',
+    icon: ArrowPathIcon,
+  },
+];
+
+export const featuredPost = {
+  body: 'Integer id nunc sit semper purus. Bibendum at lacus ut arcu blandit montes vitae auctor libero. Hac condimentum dignissim nibh vulputate ut nunc. Amet nibh orci mi venenatis blandit vel et proin. Non hendrerit in vel ac diam.',
+  author: {
+    name: 'Brenna Goyette',
+    handle: 'brennagoyette',
+    imageUrl:
+      'https://images.unsplash.com/photo-1550525811-e5869dd03032?ixlib=rb-=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=1024&h=1024&q=80',
+    logoUrl: 'https://tailwindui.com/img/logos/savvycal-logo-gray-900.svg',
+  },
+};
+
+export const recentPosts = [
+  [
+    [
+      {
+        body: 'Laborum quis quam. Dolorum et ut quod quia. Voluptas numquam delectus nihil. Aut enim doloremque et ipsam.',
+        author: {
+          name: 'Leslie Alexander',
+          handle: 'lesliealexander',
+          imageUrl:
+            'https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+        },
+      },
+      // More testimonials...
+    ],
+    [
+      {
+        body: 'Aut reprehenderit voluptatem eum asperiores beatae id. Iure molestiae ipsam ut officia rem nulla blanditiis.',
+        author: {
+          name: 'Lindsay Walton',
+          handle: 'lindsaywalton',
+          imageUrl:
+            'https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+        },
+      },
+      // More testimonials...
+    ],
+  ],
+  [
+    [
+      {
+        body: 'Voluptas quos itaque ipsam in voluptatem est. Iste eos blanditiis repudiandae. Earum deserunt enim molestiae ipsum perferendis recusandae saepe corrupti.',
+        author: {
+          name: 'Tom Cook',
+          handle: 'tomcook',
+          imageUrl:
+            'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+        },
+      },
+      // More testimonials...
+    ],
+    [
+      {
+        body: 'Molestias ea earum quos nostrum doloremque sed. Quaerat quasi aut velit incidunt excepturi rerum voluptatem minus harum.',
+        author: {
+          name: 'Leonard Krasner',
+          handle: 'leonardkrasner',
+          imageUrl:
+            'https://images.unsplash.com/photo-1519345182560-3f2917c472ef?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+        },
+      },
+      // More testimonials...
+    ],
+  ],
+];
+
+export const footerNavigation = {
+  solutions: [
+    { name: 'Marketing', href: '#' },
+    { name: 'Analytics', href: '#' },
+    { name: 'Commerce', href: '#' },
+    { name: 'Insights', href: '#' },
+  ],
+  support: [
+    { name: 'Pricing', href: '#' },
+    { name: 'Documentation', href: '#' },
+    { name: 'Guides', href: '#' },
+    { name: 'API Status', href: '#' },
+  ],
+  company: [
+    { name: 'About', href: '#' },
+    { name: 'Blog', href: '#' },
+    { name: 'Jobs', href: '#' },
+    { name: 'Press', href: '#' },
+    { name: 'Partners', href: '#' },
+  ],
+  legal: [
+    { name: 'Claim', href: '#' },
+    { name: 'Privacy', href: '#' },
+    { name: 'Terms', href: '#' },
+  ],
+  social: [
+    {
+      name: 'Facebook',
+      href: '#',
+    },
+    {
+      name: 'Instagram',
+      href: '#',
+    },
+    {
+      name: 'Twitter',
+      href: '#',
+    },
+    {
+      name: 'GitHub',
+      href: '#',
+    },
+    {
+      name: 'YouTube',
+      href: '#',
+    },
+  ] as { name: SocialVariant; href: string }[],
+};
