+++ libs/waweb-ui/src/ui/views/home-view/home-view.tsx
@@ -1,0 +1,793 @@
+import { Dialog, Disclosure, Popover, Transition } from '@headlessui/react';
+import { ChevronDownIcon, ChevronRightIcon } from '@heroicons/react/20/solid';
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline';
+import { Container, Heading, Text } from '@watheia/base-ui';
+import clsx from 'clsx';
+import Image from 'next/image';
+import Link from 'next/link';
+import { Fragment, useState } from 'react';
+import { SocialIcon } from '../../social-icon/social-icon';
+import {
+  callsToAction,
+  featuredPost,
+  features,
+  footerNavigation,
+  highlights,
+  recentPosts,
+  solutions,
+} from './home-view.fixtures';
+import { HomeViewProps } from './home-view.types';
+
+const Logo = () => (
+  <Link href="#" className="-m-1.5 p-1.5">
+    <span className="sr-only">Watheia Labs</span>
+    <Image
+      className="h-8 w-auto"
+      src="/images/icon.png"
+      width={300}
+      height={106}
+      alt=""
+      priority
+    />
+  </Link>
+);
+
+const Header = () => {
+  const [mobileMenuOpen, setMobileMenuOpen] = useState<boolean>(false);
+  return (
+    <header className="relative isolate z-10">
+      <nav
+        className="mx-auto flex max-w-7xl items-center justify-between p-6 lg:px-8"
+        aria-label="Global"
+      >
+        <div className="flex pr-12">
+          <Logo />
+        </div>
+        <div className="flex lg:hidden">
+          <button
+            type="button"
+            className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-neutral-700"
+            onClick={() => setMobileMenuOpen(true)}
+          >
+            <span className="sr-only">Open main menu</span>
+            <Bars3Icon className="h-6 w-6" aria-hidden="true" />
+          </button>
+        </div>
+        <Popover.Group className="hidden lg:flex lg:gap-x-12">
+          <Popover>
+            <Popover.Button className="flex items-center gap-x-1 text-sm font-semibold leading-6 text-neutral-900">
+              Solutions
+              <ChevronDownIcon
+                className="h-5 w-5 flex-none text-neutral-400"
+                aria-hidden="true"
+              />
+            </Popover.Button>
+
+            <Transition
+              as={Fragment}
+              enter="transition ease-out duration-200"
+              enterFrom="opacity-0 -translate-y-1"
+              enterTo="opacity-100 translate-y-0"
+              leave="transition ease-in duration-150"
+              leaveFrom="opacity-100 translate-y-0"
+              leaveTo="opacity-0 -translate-y-1"
+            >
+              <Popover.Panel className="absolute inset-x-0 top-0 -z-10 bg-white pt-14 shadow-lg ring-1 ring-neutral-900/5">
+                <div className="mx-auto grid max-w-7xl grid-cols-4 gap-x-4 px-6 py-10 lg:px-8 xl:gap-x-8">
+                  {solutions.map((item) => (
+                    <div
+                      key={item.name}
+                      className="group relative rounded-lg p-6 text-sm leading-6 hover:bg-neutral-50"
+                    >
+                      <div className="flex h-11 w-11 items-center justify-center rounded-lg bg-neutral-50 group-hover:bg-white">
+                        <item.icon
+                          className="h-6 w-6 text-neutral-600 group-hover:text-indigo-600"
+                          aria-hidden="true"
+                        />
+                      </div>
+                      <Link
+                        href={item.href}
+                        className="mt-6 block font-semibold text-neutral-900"
+                      >
+                        {item.name}
+                        <span className="absolute inset-0" />
+                      </Link>
+                      <p className="mt-1 text-neutral-600">{item.description}</p>
+                    </div>
+                  ))}
+                </div>
+                <div className="bg-neutral-50">
+                  <div className="mx-auto max-w-7xl px-6 lg:px-8">
+                    <div className="grid grid-cols-3 divide-x divide-neutral-900/5 border-x border-neutral-900/5">
+                      {callsToAction.map((item) => (
+                        <Link
+                          key={item.name}
+                          href={item.href}
+                          className="flex items-center justify-center gap-x-2.5 p-3 text-sm font-semibold leading-6 text-neutral-900 hover:bg-neutral-100"
+                        >
+                          <item.icon
+                            className="h-5 w-5 flex-none text-neutral-400"
+                            aria-hidden="true"
+                          />
+                          {item.name}
+                        </Link>
+                      ))}
+                    </div>
+                  </div>
+                </div>
+              </Popover.Panel>
+            </Transition>
+          </Popover>
+
+          <Link href="#" className="text-sm font-semibold leading-6 text-neutral-900">
+            Demo
+          </Link>
+          <Link href="#" className="text-sm font-semibold leading-6 text-neutral-900">
+            Blog
+          </Link>
+          <Link href="#" className="text-sm font-semibold leading-6 text-neutral-900">
+            Contact
+          </Link>
+        </Popover.Group>
+        <div className="hidden lg:flex lg:flex-1 lg:justify-end">
+          <Link href="/signin" className="text-sm font-semibold leading-6 text-neutral-900">
+            Log in <span aria-hidden="true">&rarr;</span>
+          </Link>
+        </div>
+      </nav>
+      <Dialog as="div" className="lg:hidden" open={mobileMenuOpen} onClose={setMobileMenuOpen}>
+        <div className="fixed inset-0 z-10" />
+        <Dialog.Panel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-neutral-900/10">
+          <div className="flex items-center justify-between">
+            <Logo />
+            <button
+              type="button"
+              className="-m-2.5 rounded-md p-2.5 text-neutral-700"
+              onClick={() => setMobileMenuOpen(false)}
+            >
+              <span className="sr-only">Close menu</span>
+              <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+            </button>
+          </div>
+          <div className="mt-6 flow-root">
+            <div className="-my-6 divide-y divide-neutral-500/10">
+              <div className="space-y-2 py-6">
+                <Disclosure as="div" className="-mx-3">
+                  {({ open }: { open: boolean }) => (
+                    <>
+                      <Disclosure.Button className="flex w-full items-center justify-between rounded-lg py-2 pl-3 pr-3.5 text-base font-semibold leading-7 text-neutral-900 hover:bg-neutral-50">
+                        Solutions
+                        <ChevronDownIcon
+                          className={clsx(open ? 'rotate-180' : '', 'h-5 w-5 flex-none')}
+                          aria-hidden="true"
+                        />
+                      </Disclosure.Button>
+                      <Disclosure.Panel className="mt-2 space-y-2">
+                        {[...solutions, ...callsToAction].map((item) => (
+                          <Disclosure.Button
+                            key={item.name}
+                            as={Link}
+                            href={item.href}
+                            className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-neutral-900 hover:bg-neutral-50"
+                          >
+                            {item.name}
+                          </Disclosure.Button>
+                        ))}
+                      </Disclosure.Panel>
+                    </>
+                  )}
+                </Disclosure>
+                <Link
+                  href="/blog"
+                  className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-neutral-900 hover:bg-neutral-50"
+                >
+                  Blog
+                </Link>
+                <Link
+                  href="/demo"
+                  className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-neutral-900 hover:bg-neutral-50"
+                >
+                  Demo
+                </Link>
+                <Link
+                  href="/contact"
+                  className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-neutral-900 hover:bg-neutral-50"
+                >
+                  Contact
+                </Link>
+              </div>
+              <div className="py-6">
+                <Link
+                  href="/signin"
+                  className="-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-neutral-900 hover:bg-neutral-50"
+                >
+                  Log in
+                </Link>
+              </div>
+            </div>
+          </div>
+        </Dialog.Panel>
+      </Dialog>
+    </header>
+  );
+};
+
+const Hero = () => (
+  <div className="relative isolate pt-14">
+    <svg
+      className="absolute inset-0 -z-10 h-full w-full stroke-neutral-200 [mask-image:radial-gradient(100%_100%_at_top_right,white,transparent)]"
+      aria-hidden="true"
+    >
+      <defs>
+        <pattern
+          id="83fd4e5a-9d52-42fc-97b6-718e5d7ee527"
+          width={200}
+          height={200}
+          x="50%"
+          y={-1}
+          patternUnits="userSpaceOnUse"
+        >
+          <path d="M100 200V.5M.5 .5H200" fill="none" />
+        </pattern>
+      </defs>
+      <svg x="50%" y={-1} className="overflow-visible fill-neutral-50">
+        <path
+          d="M-100.5 0h201v201h-201Z M699.5 0h201v201h-201Z M499.5 400h201v201h-201Z M-300.5 600h201v201h-201Z"
+          strokeWidth={0}
+        />
+      </svg>
+      <rect
+        width="100%"
+        height="100%"
+        strokeWidth={0}
+        fill="url(#83fd4e5a-9d52-42fc-97b6-718e5d7ee527)"
+      />
+    </svg>
+    <div className="mx-auto max-w-7xl px-6 py-24 sm:py-32 lg:flex lg:items-center lg:gap-x-10 lg:px-8 lg:py-40">
+      <div className="mx-auto max-w-2xl lg:mx-0 lg:flex-auto">
+        <div className="flex">
+          <div className="relative flex items-center gap-x-4 rounded-full px-4 py-1 text-sm leading-6 text-neutral-600 ring-1 ring-neutral-900/10 hover:ring-neutral-900/20">
+            <span className="font-semibold text-primary-600">Hire us!</span>
+            <span className="h-4 w-px bg-neutral-900/10" aria-hidden="true" />
+            <Link href="#" className="flex items-center gap-x-1">
+              <span className="absolute inset-0" aria-hidden="true" />
+              Get free consultation
+              <ChevronRightIcon className="-mr-2 h-5 w-5 text-neutral-400" aria-hidden="true" />
+            </Link>
+          </div>
+        </div>
+        <Heading level={1}>We make technology easy</Heading>
+        <Text className="mt-6 text-lg leading-8 text-neutral-600">
+          Esse id magna consectetur fugiat non dolor in ad laboris magna laborum ea consequat.
+          Nisi irure aliquip nisi adipisicing veniam voluptate id. In veniam incididunt ex
+          veniam adipisicing sit.
+        </Text>
+        <div className="mt-10 flex items-center gap-x-6">
+          <Link
+            href="#"
+            className="rounded-md bg-primary-600 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-primary-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary-600"
+          >
+            Get started
+          </Link>
+          <Link href="#" className="text-sm font-semibold leading-6 text-neutral-900">
+            Learn more <span aria-hidden="true">→</span>
+          </Link>
+        </div>
+      </div>
+      <div className="mt-16 sm:mt-24 lg:mt-0 lg:flex-shrink-0 lg:flex-grow">
+        <svg
+          viewBox="0 0 366 729"
+          role="img"
+          className="mx-auto w-[22.875rem] max-w-full drop-shadow-xl"
+        >
+          <title>App screenshot</title>
+          <defs>
+            <clipPath id="2ade4387-9c63-4fc4-b754-10e687a0d332">
+              <rect width={316} height={684} rx={36} />
+            </clipPath>
+          </defs>
+          <path
+            fill="#4B5563"
+            d="M363.315 64.213C363.315 22.99 341.312 1 300.092 1H66.751C25.53 1 3.528 22.99 3.528 64.213v44.68l-.857.143A2 2 0 0 0 1 111.009v24.611a2 2 0 0 0 1.671 1.973l.95.158a2.26 2.26 0 0 1-.093.236v26.173c.212.1.398.296.541.643l-1.398.233A2 2 0 0 0 1 167.009v47.611a2 2 0 0 0 1.671 1.973l1.368.228c-.139.319-.314.533-.511.653v16.637c.221.104.414.313.56.689l-1.417.236A2 2 0 0 0 1 237.009v47.611a2 2 0 0 0 1.671 1.973l1.347.225c-.135.294-.302.493-.49.607v377.681c0 41.213 22 63.208 63.223 63.208h95.074c.947-.504 2.717-.843 4.745-.843l.141.001h.194l.086-.001 33.704.005c1.849.043 3.442.37 4.323.838h95.074c41.222 0 63.223-21.999 63.223-63.212v-394.63c-.259-.275-.48-.796-.63-1.47l-.011-.133 1.655-.276A2 2 0 0 0 366 266.62v-77.611a2 2 0 0 0-1.671-1.973l-1.712-.285c.148-.839.396-1.491.698-1.811V64.213Z"
+          />
+          <path
+            fill="#343E4E"
+            d="M16 59c0-23.748 19.252-43 43-43h246c23.748 0 43 19.252 43 43v615c0 23.196-18.804 42-42 42H58c-23.196 0-42-18.804-42-42V59Z"
+          />
+          <foreignObject
+            width={316}
+            height={684}
+            transform="translate(24 24)"
+            clipPath="url(#2ade4387-9c63-4fc4-b754-10e687a0d332)"
+          >
+            <img
+              src="https://tailwindui.com/img/component-images/mobile-app-screenshot.png"
+              alt=""
+            />
+          </foreignObject>
+        </svg>
+      </div>
+    </div>
+  </div>
+);
+
+const LogoCloud = () => (
+  <Container>
+    <div className="mx-auto grid max-w-lg grid-cols-4 items-center gap-x-8 gap-y-12 sm:max-w-xl sm:grid-cols-6 sm:gap-x-10 sm:gap-y-14 lg:mx-0 lg:max-w-none lg:grid-cols-5">
+      <img
+        className="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
+        src="https://tailwindui.com/img/logos/158x48/transistor-logo-gray-400.svg"
+        alt="Transistor"
+        width={158}
+        height={48}
+      />
+      <img
+        className="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
+        src="https://tailwindui.com/img/logos/158x48/reform-logo-gray-400.svg"
+        alt="Reform"
+        width={158}
+        height={48}
+      />
+      <img
+        className="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
+        src="https://tailwindui.com/img/logos/158x48/tuple-logo-gray-400.svg"
+        alt="Tuple"
+        width={158}
+        height={48}
+      />
+      <img
+        className="col-span-2 max-h-12 w-full object-contain sm:col-start-2 lg:col-span-1"
+        src="https://tailwindui.com/img/logos/158x48/savvycal-logo-gray-400.svg"
+        alt="SavvyCal"
+        width={158}
+        height={48}
+      />
+      <img
+        className="col-span-2 col-start-2 max-h-12 w-full object-contain sm:col-start-auto lg:col-span-1"
+        src="https://tailwindui.com/img/logos/158x48/statamic-logo-gray-400.svg"
+        alt="Statamic"
+        width={158}
+        height={48}
+      />
+    </div>
+  </Container>
+);
+
+const Highlights = () => (
+  <Container className="pb-16 pt-20 sm:pb-24 sm:pt-32">
+    <div className="relative isolate overflow-hidden bg-neutral-900 px-6 py-20 sm:rounded-3xl sm:px-10 sm:py-24 lg:py-24 xl:px-24">
+      <div className="mx-auto grid max-w-2xl grid-cols-1 gap-x-8 gap-y-16 sm:gap-y-20 lg:mx-0 lg:max-w-none lg:grid-cols-2 lg:items-center lg:gap-y-0">
+        <div className="lg:row-start-2 lg:max-w-md">
+          <h2 className="text-3xl font-bold tracking-tight text-white sm:text-4xl">
+            Boost your productivity.
+            <br />
+            Start using our app today.
+          </h2>
+          <p className="mt-6 text-lg leading-8 text-neutral-300">
+            Ac euismod vel sit maecenas id pellentesque eu sed consectetur. Malesuada adipiscing
+            sagittis vel nulla. Ac euismod vel sit maecenas.
+          </p>
+        </div>
+        <img
+          src="https://tailwindui.com/img/component-images/dark-project-app-screenshot.png"
+          alt="Product screenshot"
+          className="relative -z-20 min-w-full max-w-xl rounded-xl shadow-xl ring-1 ring-white/10 lg:row-span-4 lg:w-[64rem] lg:max-w-none"
+          width={2432}
+          height={1442}
+        />
+        <div className="max-w-xl lg:row-start-3 lg:mt-10 lg:max-w-md lg:border-t lg:border-white/10 lg:pt-10">
+          <dl className="max-w-xl space-y-8 text-base leading-7 text-neutral-300 lg:max-w-none">
+            {highlights.map((feature) => (
+              <div key={feature.name} className="relative">
+                <dt className="ml-9 inline-block font-semibold text-white">
+                  <feature.icon
+                    className="absolute left-1 top-1 h-5 w-5 text-primary-500"
+                    aria-hidden="true"
+                  />
+                  {feature.name}
+                </dt>{' '}
+                <dd className="inline">{feature.description}</dd>
+              </div>
+            ))}
+          </dl>
+        </div>
+      </div>
+      <div
+        className="pointer-events-none absolute left-12 top-1/2 -z-10 -translate-y-1/2 transform-gpu blur-3xl lg:bottom-[-12rem] lg:top-auto lg:translate-y-0 lg:transform-gpu"
+        aria-hidden="true"
+      >
+        <div
+          className="aspect-[1155/678] w-[72.1875rem] bg-gradient-to-tr from-[#7dd3fc] to-[#0369a1] opacity-25"
+          style={{
+            clipPath:
+              'polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)',
+          }}
+        />
+      </div>
+    </div>
+  </Container>
+);
+
+const Features = () => (
+  <div className="border-t border-b border-neutral-200 bg-neutral-100 py-20 sm:py-32">
+    <Container>
+      <div className="mx-auto max-w-2xl lg:text-center">
+        <h2 className="text-base font-semibold leading-7 text-primary-600">Deploy faster</h2>
+        <p className="mt-2 text-3xl font-bold tracking-tight text-neutral-900 sm:text-4xl">
+          Everything you need to deploy your app
+        </p>
+        <p className="mt-6 text-lg leading-8 text-neutral-600">
+          Quis tellus eget adipiscing convallis sit sit eget aliquet quis. Suspendisse eget
+          egestas a elementum pulvinar et feugiat blandit at. In mi viverra elit nunc.
+        </p>
+      </div>
+      <div className="mx-auto mt-16 max-w-2xl sm:mt-20 lg:mt-24 lg:max-w-none">
+        <dl className="grid max-w-xl grid-cols-1 gap-x-8 gap-y-16 lg:max-w-none lg:grid-cols-3">
+          {features.map((feature) => (
+            <div key={feature.name} className="flex flex-col">
+              <dt className="flex items-center gap-x-3 text-base font-semibold leading-7 text-neutral-900">
+                <feature.icon
+                  className="h-5 w-5 flex-none text-primary-600"
+                  aria-hidden="true"
+                />
+                {feature.name}
+              </dt>
+              <dd className="mt-4 flex flex-auto flex-col text-base leading-7 text-neutral-600">
+                <p className="flex-auto">{feature.description}</p>
+                <p className="mt-6">
+                  <Link
+                    href={feature.href}
+                    className="text-sm font-semibold leading-6 text-primary-600"
+                  >
+                    Learn more <span aria-hidden="true">→</span>
+                  </Link>
+                </p>
+              </dd>
+            </div>
+          ))}
+        </dl>
+      </div>
+    </Container>
+  </div>
+);
+
+const Newsletter = () => (
+  <Container className="mt-32 sm:mt-56">
+    <div className="relative isolate overflow-hidden bg-neutral-900 px-6 py-24 shadow-2xl sm:rounded-3xl sm:px-24 xl:py-32">
+      <h2 className="mx-auto max-w-2xl text-center text-3xl font-bold tracking-tight text-white sm:text-4xl">
+        Get notified when we’re launching.
+      </h2>
+      <p className="mx-auto mt-2 max-w-xl text-center text-lg leading-8 text-neutral-300">
+        Reprehenderit ad esse et non officia in nulla. Id proident tempor incididunt nostrud
+        nulla et culpa.
+      </p>
+      <form className="mx-auto mt-10 flex max-w-md gap-x-4">
+        <label htmlFor="email-address" className="sr-only">
+          Email address
+        </label>
+        <input
+          id="email-address"
+          name="email"
+          type="email"
+          autoComplete="email"
+          required
+          className="min-w-0 flex-auto rounded-md border-0 bg-white/5 px-3.5 py-2 text-white shadow-sm ring-1 ring-inset ring-white/10 focus:ring-2 focus:ring-inset focus:ring-white sm:text-sm sm:leading-6"
+          placeholder="Enter your email"
+        />
+        <button
+          type="submit"
+          className="flex-none rounded-md bg-white px-3.5 py-2.5 text-sm font-semibold text-neutral-900 shadow-sm hover:bg-neutral-100 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white"
+        >
+          Notify me
+        </button>
+      </form>
+      <svg
+        viewBox="0 0 1024 1024"
+        className="absolute left-1/2 top-1/2 -z-10 h-[64rem] w-[64rem] -translate-x-1/2"
+        aria-hidden="true"
+      >
+        <circle
+          cx={512}
+          cy={512}
+          r={512}
+          fill="url(#759c1415-0410-454c-8f7c-9a820de03641)"
+          fillOpacity="0.7"
+        />
+        <defs>
+          <radialGradient
+            id="759c1415-0410-454c-8f7c-9a820de03641"
+            cx={0}
+            cy={0}
+            r={1}
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(512 512) rotate(90) scale(512)"
+          >
+            <stop stopColor="#7dd3fc" />
+            <stop offset={1} stopColor="#0369a1" stopOpacity={0} />
+          </radialGradient>
+        </defs>
+      </svg>
+    </div>
+  </Container>
+);
+
+const Testimonials = () => (
+  <div className="relative isolate mt-32 sm:mt-56 sm:pt-32">
+    <svg
+      className="absolute inset-0 -z-10 hidden h-full w-full stroke-neutral-200 [mask-image:radial-gradient(64rem_64rem_at_top,white,transparent)] sm:block"
+      aria-hidden="true"
+    >
+      <defs>
+        <pattern
+          id="55d3d46d-692e-45f2-becd-d8bdc9344f45"
+          width={200}
+          height={200}
+          x="50%"
+          y={0}
+          patternUnits="userSpaceOnUse"
+        >
+          <path d="M.5 200V.5H200" fill="none" />
+        </pattern>
+      </defs>
+      <svg x="50%" y={0} className="overflow-visible fill-neutral-50">
+        <path
+          d="M-200.5 0h201v201h-201Z M599.5 0h201v201h-201Z M399.5 400h201v201h-201Z M-400.5 600h201v201h-201Z"
+          strokeWidth={0}
+        />
+      </svg>
+      <rect
+        width="100%"
+        height="100%"
+        strokeWidth={0}
+        fill="url(#55d3d46d-692e-45f2-becd-d8bdc9344f45)"
+      />
+    </svg>
+    <div className="relative">
+      <div
+        className="absolute inset-x-0 top-1/2 -z-10 -translate-y-1/2 transform-gpu overflow-hidden opacity-30 blur-3xl"
+        aria-hidden="true"
+      >
+        <div
+          className="ml-[max(50%,38rem)] aspect-[1313/771] w-[82.0625rem] bg-gradient-to-tr from-[#7dd3fc] to-[#0369a1]"
+          style={{
+            clipPath:
+              'polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)',
+          }}
+        />
+      </div>
+      <div
+        className="absolute inset-x-0 top-0 -z-10 flex transform-gpu overflow-hidden pt-8 opacity-25 blur-3xl xl:justify-end"
+        aria-hidden="true"
+      >
+        <div
+          className="ml-[-22rem] aspect-[1313/771] w-[82.0625rem] flex-none origin-top-right rotate-[30deg] bg-gradient-to-tr from-[#7dd3fc] to-[#0369a1] xl:ml-0 xl:mr-[calc(50%-12rem)]"
+          style={{
+            clipPath:
+              'polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)',
+          }}
+        />
+      </div>
+      <div className="mx-auto max-w-7xl px-6 lg:px-8">
+        <div className="mx-auto max-w-xl sm:text-center">
+          <h2 className="text-lg font-semibold leading-8 tracking-tight text-primary-600">
+            Testimonials
+          </h2>
+          <p className="mt-2 text-3xl font-bold tracking-tight text-neutral-900 sm:text-4xl">
+            We have worked with thousands of amazing people
+          </p>
+        </div>
+        <div className="mx-auto mt-16 grid max-w-2xl grid-cols-1 grid-rows-1 gap-8 text-sm leading-6 text-neutral-900 sm:mt-20 sm:grid-cols-2 xl:mx-0 xl:max-w-none xl:grid-flow-col xl:grid-cols-4">
+          <figure className="col-span-2 hidden sm:block sm:rounded-2xl sm:bg-white sm:shadow-lg sm:ring-1 sm:ring-neutral-900/5 xl:col-start-2 xl:row-end-1">
+            <blockquote className="p-12 text-xl font-semibold leading-8 tracking-tight text-neutral-900">
+              <p>{`“${featuredPost.body}”`}</p>
+            </blockquote>
+            <figcaption className="flex items-center gap-x-4 border-t border-neutral-900/10 px-6 py-4">
+              <img
+                className="h-10 w-10 flex-none rounded-full bg-neutral-50"
+                src={featuredPost.author.imageUrl}
+                alt=""
+              />
+              <div className="flex-auto">
+                <div className="font-semibold">{featuredPost.author.name}</div>
+                <div className="text-neutral-600">{`@${featuredPost.author.handle}`}</div>
+              </div>
+              <img className="h-10 w-auto flex-none" src={featuredPost.author.logoUrl} alt="" />
+            </figcaption>
+          </figure>
+          {recentPosts.map((columnGroup, columnGroupIdx) => (
+            <div key={columnGroupIdx} className="space-y-8 xl:contents xl:space-y-0">
+              {columnGroup.map((column, columnIdx) => (
+                <div
+                  key={columnIdx}
+                  className={clsx(
+                    (columnGroupIdx === 0 && columnIdx === 0) ||
+                      (columnGroupIdx === recentPosts.length - 1 &&
+                        columnIdx === columnGroup.length - 1)
+                      ? 'xl:row-span-2'
+                      : 'xl:row-start-1',
+                    'space-y-8'
+                  )}
+                >
+                  {column.map((testimonial) => (
+                    <figure
+                      key={testimonial.author.handle}
+                      className="rounded-2xl bg-white p-6 shadow-lg ring-1 ring-neutral-900/5"
+                    >
+                      <blockquote className="text-neutral-900">
+                        <p>{`“${testimonial.body}”`}</p>
+                      </blockquote>
+                      <figcaption className="mt-6 flex items-center gap-x-4">
+                        <img
+                          className="h-10 w-10 rounded-full bg-neutral-50"
+                          src={testimonial.author.imageUrl}
+                          alt=""
+                        />
+                        <div>
+                          <div className="font-semibold">{testimonial.author.name}</div>
+                          <div className="text-neutral-600">{`@${testimonial.author.handle}`}</div>
+                        </div>
+                      </figcaption>
+                    </figure>
+                  ))}
+                </div>
+              ))}
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  </div>
+);
+
+const Footer = () => (
+  <footer className="mt-32 bg-neutral-900 sm:mt-56" aria-labelledby="footer-heading">
+    <h2 id="footer-heading" className="sr-only">
+      Footer
+    </h2>
+    <Container className="px-6 pb-8 pt-16 sm:pt-24 lg:px-8 lg:pt-32">
+      <div className="xl:grid xl:grid-cols-3 xl:gap-8">
+        <Image
+          src="/images/logo-alt.png"
+          alt="Watheia Labs"
+          width={301}
+          height={107}
+          quality={100}
+        />
+        <div className="mt-16 grid grid-cols-2 gap-8 xl:col-span-2 xl:mt-0">
+          <div className="md:grid md:grid-cols-2 md:gap-8">
+            <div>
+              <h3 className="text-sm font-semibold leading-6 text-white">Solutions</h3>
+              <ul role="list" className="mt-6 space-y-4">
+                {footerNavigation.solutions.map((item) => (
+                  <li key={item.name}>
+                    <Link
+                      href={item.href}
+                      className="text-sm leading-6 text-neutral-300 hover:text-white"
+                    >
+                      {item.name}
+                    </Link>
+                  </li>
+                ))}
+              </ul>
+            </div>
+            <div className="mt-10 md:mt-0">
+              <h3 className="text-sm font-semibold leading-6 text-white">Support</h3>
+              <ul role="list" className="mt-6 space-y-4">
+                {footerNavigation.support.map((item) => (
+                  <li key={item.name}>
+                    <Link
+                      href={item.href}
+                      className="text-sm leading-6 text-neutral-300 hover:text-white"
+                    >
+                      {item.name}
+                    </Link>
+                  </li>
+                ))}
+              </ul>
+            </div>
+          </div>
+          <div className="md:grid md:grid-cols-2 md:gap-8">
+            <div>
+              <h3 className="text-sm font-semibold leading-6 text-white">Company</h3>
+              <ul role="list" className="mt-6 space-y-4">
+                {footerNavigation.company.map((item) => (
+                  <li key={item.name}>
+                    <Link
+                      href={item.href}
+                      className="text-sm leading-6 text-neutral-300 hover:text-white"
+                    >
+                      {item.name}
+                    </Link>
+                  </li>
+                ))}
+              </ul>
+            </div>
+            <div className="mt-10 md:mt-0">
+              <h3 className="text-sm font-semibold leading-6 text-white">Legal</h3>
+              <ul role="list" className="mt-6 space-y-4">
+                {footerNavigation.legal.map((item) => (
+                  <li key={item.name}>
+                    <Link
+                      href={item.href}
+                      className="text-sm leading-6 text-neutral-300 hover:text-white"
+                    >
+                      {item.name}
+                    </Link>
+                  </li>
+                ))}
+              </ul>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div className="mt-16 border-t border-white/10 pt-8 sm:mt-20 lg:mt-24 lg:flex lg:items-center lg:justify-between">
+        <div>
+          <h3 className="text-sm font-semibold leading-6 text-white">
+            Subscribe to our newsletter
+          </h3>
+          <p className="mt-2 text-sm leading-6 text-neutral-300">
+            The latest news, articles, and resources, sent to your inbox weekly.
+          </p>
+        </div>
+        <form className="mt-6 sm:flex sm:max-w-md lg:mt-0">
+          <label htmlFor="email-address" className="sr-only">
+            Email address
+          </label>
+          <input
+            type="email"
+            name="email-address"
+            id="email-address"
+            autoComplete="email"
+            required
+            className="w-full min-w-0 appearance-none rounded-md border-0 bg-white/5 px-3 py-1.5 text-base text-white shadow-sm ring-1 ring-inset ring-white/10 placeholder:text-neutral-500 focus:ring-2 focus:ring-inset focus:ring-primary-500 sm:w-56 sm:text-sm sm:leading-6"
+            placeholder="Enter your email"
+          />
+          <div className="mt-4 sm:ml-4 sm:mt-0 sm:flex-shrink-0">
+            <button
+              type="submit"
+              className="flex w-full items-center justify-center rounded-md bg-primary-500 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-primary-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary-500"
+            >
+              Subscribe
+            </button>
+          </div>
+        </form>
+      </div>
+      <div className="mt-8 border-t border-white/10 pt-8 md:flex md:items-center md:justify-between">
+        <div className="flex space-x-6 md:order-2">
+          {footerNavigation.social.map((item) => (
+            <Link
+              key={item.name}
+              href={item.href}
+              className="text-neutral-500 hover:text-neutral-400"
+            >
+              <span className="sr-only">{item.name}</span>
+              <SocialIcon variant={item.name} className="h-6 w-6" aria-hidden="true" />
+            </Link>
+          ))}
+        </div>
+        <p className="mt-8 text-xs leading-5 text-neutral-400 md:order-1 md:mt-0">
+          &copy; {new Date().getFullYear()} Watheia Labs, LLC. All rights reserved.
+        </p>
+      </div>
+    </Container>
+  </footer>
+);
+
+export function HomeView({ featuredPost, recentPosts, ...props }: HomeViewProps) {
+  return (
+    <article {...props}>
+      <Header />
+      <main>
+        <Hero />
+        <LogoCloud />
+        <Highlights />
+        <Features />
+        <Newsletter />
+        <Testimonials />
+      </main>
+      <Footer />
+    </article>
+  );
+}
+
+export default HomeView;
