+++ libs/waweb-ui/src/ui/views/home-view/index.ts
@@ -1,0 +1,2 @@
+export { HomeView, default } from './home-view';
+export type { HomeViewProps } from './home-view.types';
